Pipeline mit besseren Variablennamen Stanford wird aufgerufen, wenn ReVerb mit der Frage nichts anfangen kann. Das ist viel zu häufig geschehen. process_corpus: auf exit status wird reagiert, wenn corpus nicht existiert Process_question bekommt Dateiname mit Frage übergeben

+++ process_corpus.sh
@@ -14,6 +14,11 @@ fi
 corpus="$1"
 db="$2"
 
+if [ ! -f "$corpus" ]; then
+	>&2 echo "Das Dokument $corpus existiert nicht!"
+	exit 1
+fi
+
 java -Xmx512m -jar ReVerb/reverb-latest.jar "$corpus" > $reverbout.txt 2> /dev/null
 echo "filename;sentence number;arg1;rel;arg2;arg1 start;arg1 end;rel start;rel end;arg2 start;arg2 end;conf;sentence words;sentence pos tags;sentence chunk tags;arg1 normalized;rel normalized;arg2 normalized" > $reverbout.csv
 sed 's/	/;/g' $reverbout.txt >> $reverbout.csv
